Art Week 2 JM
Pushing Hunter and Player Models, the textures for them have been implemented and made into materials of both 2k and 4k quality. Furthermore some animations from Mixamo have been added in and named for both.

NOTE: Should be able to drop in the T-Pose (as it has the mixamo rig attached), drop the material onto it, and use that single object as the animator with the animation infos that exist in the other animation assets